Creating a Measure element to make things easier.
